fix: ignore stderr messages
Since we're running on a tty, our parent process isn't the display server (which handles the stderr messages gracefully); our parent is a shell inside our tty, which shows the kernel panic errors in case one appears.

When a kernel panic error occurs, the display screen messes up, and the UI breaks.

Therefore, if the errors break the UI and the program doesn't handle them, ignore them. If the user wants to know why something isn't working as expected, they can switch to another tty and check the logs.